+++ src/scheduler/admin.py
@@ -2,11 +2,13 @@
 from __future__ import unicode_literals
 
 from django.contrib import admin
-from models import ScheduledAction
+from models import ScheduledEmailAction
 
 
-class ScheduledActionAdmin(admin.ModelAdmin):
-    list_display = ('user', 'workflow', 'type', 'created', 'executed',
-                    'status', 'payload')
+class ScheduledEmailActionAdmin(admin.ModelAdmin):
+    list_display = ('user', 'type', 'created', 'execute',
+                    'status', 'action', 'subject', 'email_column',
+                    'send_confirmation', 'track_read', 'add_column',
+                    'message')
 
-admin.site.register(ScheduledAction, ScheduledActionAdmin)
+admin.site.register(ScheduledEmailAction, ScheduledEmailActionAdmin)